projects/hallucination-resistant-rag/requirements.txt
pdfplumber
sentence-transformers
numpy
faiss-cpu
fastapi
uvicorn
python-multipart
requests
